feat: Added error posters images for main view, add onClickListener if poster doesn't load
Now if user clicks on a poster that didn't load correctly and displays Picasso's error poster
it will take him/her to the movie details screen with the corresponding data. Each error loading
poster image corresponds to the right context. One for the main screen and one for the details screen.